requirements.txt
pushbullet.py
selenium
beautifulsoup4